Update deck.c
AssignCards: This function now assigns cards to all players, and also assigns small blind and big blind to the players if the gamestate is in preflop. If the gamestate is not preflop, the function moves the current small and big blinds to the next players. 
PERFLOP1: This is a new function, which runs the game. The function keeps running until all the bids are equal. when the program will stop and print out the cards and bets of each player. If a player folds, it will not prompt that player to input again. 
EQUALBIDS: This function checks if the bids from all players are equal. It ignores the bets from players who have already folded. It only works if player 0 did not fold, for now, I can change that soon.
The preflop stage seems to be working well now  although there are a few things that need to be changed.