Update game.c
Added new functions: CheckPlayer() checks a player's deck and assigns him a number according to his sequence and matches.
Sequence_Winner() Compares the ranking of different players and assigns a winner
The IsSequence() Functions are still to be implemented. IsFlush works for now.
SortByRank and LargestinArray functions were made for isFlush to work. SortbyRank() Takes a player 's cards and arranges the cards in ascending order of rank with the community cards.